working on #206 : adding support for x64 to asm2binx and elc

+++ elenasrc2/tools/asm2bin/assembler.h
@@ -186,7 +186,7 @@ struct TokenInfo
       return 0; // dummy return
    }
 
-   int readLongInteger(Map<ident_t, ref64_t>& constants)
+   long long readLongInteger(Map<ident_t, ref64_t>& constants)
    {
       read();
       ref64_t integer;
